max(int, long int) does not exist ==> explicitly cast to int

+++ core/vil/algo/vil_correlate_1d.h
@@ -142,8 +142,8 @@ class vil_correlate_1d_resource : public vil_image_resource
                                                  unsigned j0, unsigned nj) const
   {
     if (i0 + ni > src_->ni() || j0 + nj > src_->nj())  return 0;
-    const unsigned lsrc = (unsigned)vcl_max(0,int(i0) + klo_); // lhs of input window
-    const unsigned hsrc = vcl_min(src_->ni(),i0 + ni - klo_ + khi_); // 1+rhs of input window.
+    const unsigned lsrc = (unsigned)vcl_max(0,int(i0+klo_)); // lhs of input window
+    const unsigned hsrc = vcl_min(src_->ni(),(unsigned int)(i0+ni-klo_+khi_)); // 1+rhs of input window.
     const unsigned lboundary = vcl_min((unsigned) -klo_, i0); // width of lhs boundary area.
     assert (hsrc > lsrc);
     vil_image_view_base_sptr vs = src_->get_view(lsrc, hsrc-lsrc, j0, nj);
